requirements.txt
# YouTube RAG Question Answering App

streamlit>=1.28.0
youtube-transcript-api>=1.2.2
langchain>=0.3.27
langchain-community>=0.3.29
langchain-core>=0.3.76
langchain-huggingface>=0.3.1
langchain-perplexity>=0.1.2
faiss-cpu>=1.12.0
sentence-transformers>=5.1.1
python-dotenv>=1.1.1

# Optional: for better performance
torch>=2.0.0
transformers>=4.30.0